+++ app/components/map/overlays/pilot/MapPilotOverlay.vue
@@ -324,7 +324,6 @@ const ctafFrequency = computed(() => {
 });
 
 const pilot = computed(() => props.overlay.data.pilot);
-const textCoords = computed(() => dataStore.vatsim.data.keyedPilots.value[props.overlay.data.pilot.cid.toString()]?.longitude.toString() + dataStore.vatsim.data.keyedPilots.value[props.overlay.data.pilot.cid.toString()]?.latitude);
 const airportInfo = computed(() => {
     return props.overlay.data.airport;
 });
@@ -646,22 +645,7 @@ useUpdateCallback(['short'], async () => {
     }
 });
 
-function focusOnAircraft() {
-    if (!props.overlay.data.tracked) return;
-
-    let center = [dataStore.vatsim.data.keyedPilots.value[props.overlay.data.pilot.cid.toString()]?.longitude ?? pilot.value.longitude, dataStore.vatsim.data.keyedPilots.value[props.overlay.data.pilot.cid.toString()]?.latitude ?? pilot.value.latitude];
-
-    if (ownFlight.value?.cid === props.overlay.data.pilot.cid && dataStore.vatsim.selfCoordinate.value) center = dataStore.vatsim.selfCoordinate.value.coordinate;
-
-    map.value?.getView().animate({
-        center,
-        duration: 300,
-    });
-}
-
-watch([textCoords, dataStore.vatsim.selfCoordinate], focusOnAircraft);
 watch(() => props.overlay.data.tracked, val => {
-    focusOnAircraft();
     if (val) {
         mapStore.overlays.filter(x => x.type === 'pilot' && x.data.tracked && x.key !== pilot.value.cid.toString()).forEach(x => {
             (x as StoreOverlayPilot).data.tracked = false;
@@ -691,12 +675,10 @@ onMounted(() => {
     });
 
     map.value?.on('pointerdrag', handlePointerDrag);
-    map.value?.on('moveend', focusOnAircraft);
 });
 
 onBeforeUnmount(() => {
     map.value?.un('pointerdrag', handlePointerDrag);
-    map.value?.un('moveend', focusOnAircraft);
 
     if (enrouteAircraftPath.value) {
         delete enrouteAircraftPath.value[props.overlay.key];